Move tests to test/src directory
