+++ test/src/fib-generator-test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 import { assert } from 'chai';
-import fibGenerator from '../src/fib-generator';
+import fibGenerator from '../../src/fib-generator';
 
 describe('FibGenerator', () => {
   function assertNextValueAndDone(next, expectedValue) {
